Modify original Dockerfile to only handle impex part of a numbered build
An oracle database is provisioned in some way.

This Dockerfile creates an image that will load the data from a specific
numbered build of kuali student into that oracle database.

The oracle jdbc url can either be specified through the ORACLE_DBA_URL
environment variable or it can be autodetected through the linked DB alias.
( e.g. docker run --link oracle:db this-image )

When run this images container will stop once the impexing process has completed

If the oracle instance is in a docker container this would be the time to commit
it so that it can be archived for later use.

Restarting the oracle container will be faster than reimpexing.